Add asdf & Make targets for producing a simulator binary
Because the simulator is meant to take in a design and run a simulation, it
seems appropriate to be able to generate a single executable binary for
performing these simulations. We now have the ability to do that.

Taken with inspiration from:
https://www.reddit.com/r/lisp/comments/5jbsle/comment/dbf74vt/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button
https://github.com/Shirakumo/kandria/blob/ac4473dbe3bcb614abdbbb52ebd8e657a0a1882c/kandria.asd#L3
https://github.com/Shinmera/deploy

Only Linux is supported as a binary target currently (even though the original
source code is currently platform-agnostic).
See the link below for an explanation about #+
https://stackoverflow.com/q/8651695